requirements.txt
flask
mysql-connector-python
flask-cors
scikit-learn
pandas
python-dotenv